+++ src/main/java/org/nette/latte/syntaxHighlighter/LatteColorSettingsPage.java
@@ -20,6 +20,7 @@ public class LatteColorSettingsPage implements ColorSettingsPage {
 		new AttributesDescriptor("Tag comment", LatteSyntaxHighlighter.MACRO_COMMENT),
 		new AttributesDescriptor("Block name", LatteSyntaxHighlighter.MACRO_BLOCK_NAME),
 		new AttributesDescriptor("Destination link", LatteSyntaxHighlighter.MACRO_LINK_DESTINATION),
+		new AttributesDescriptor("Destination file", LatteSyntaxHighlighter.MACRO_FILE_PATH),
 		new AttributesDescriptor("Attribute n:tag Name", LatteSyntaxHighlighter.HTML_NATTR_NAME),
 		new AttributesDescriptor("Attribute n:tag Value", LatteSyntaxHighlighter.HTML_NATTR_VALUE),
 		new AttributesDescriptor("PHP variable", LatteSyntaxHighlighter.MACRO_ARGS_VAR),